refactor(middlewares): separating middlewares into a folder for better organization
