+++ src/middlewares/errorHandlerMiddleware.js
@@ -1,4 +1,3 @@
 module.exports = function errorHandlerMiddleware(err, req, res, next) {
-  console.error(err.stack);
   res.status(400).json({ message: err.message });
 };
